requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.4
lxml==4.9.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlite3
fake-useragent==1.4.0
time
logging
json
csv
re
urllib3==2.0.7
webdriver-manager==4.0.1